im retarded its working now

+++ Assets/HealthBar.cs
@@ -6,6 +6,6 @@ public class HealthBar : MonoBehaviour {
 
 	public float HealthRemaining;
 	void Update(){
-		transform.localScale = new Vector3((float)HealthRemaining,1,1);
+		transform.localScale = new Vector3(HealthRemaining,1f,1f);
 	}
 }
